Modula-2 bootstrap parameter passing type equivalence bugfix.
This patch removes the unnecesary casts when pointers are passed
as parameters, when pointers are assigned and when constants are
assigned to designators.

gcc/m2/ChangeLog:

	* mc/decl.mod: (isZtypeEquivalent) New procedure function.
	(isEquivalentType) New procedure function.  (doExprCastC)
	Added header comment and check for type equivalence.
	(needsCast) Added header comment and skip types for
	the formal and actual parameters before comparison.
	Add check for actual = nil parameter. (checkSystemCast)
	check for actual = nil being passed to a pointer parameter.
	* mc-boot/GArgs.c: Rebuilt.
	* mc-boot/GCmdArgs.c: Rebuilt.
	* mc-boot/GDebug.c: Rebuilt.
	* mc-boot/GDynamicStrings.c: Rebuilt.
	* mc-boot/GEnvironment.c: Rebuilt.
	* mc-boot/GFIO.c: Rebuilt.
	* mc-boot/GFormatStrings.c: Rebuilt.
	* mc-boot/GIO.c: Rebuilt.
	* mc-boot/GIndexing.c: Rebuilt.
	* mc-boot/GM2RTS.c: Rebuilt.
	* mc-boot/GNumberIO.c: Rebuilt.
	* mc-boot/GPushBackInput.c: Rebuilt.
	* mc-boot/GRTExceptions.c: Rebuilt.
	* mc-boot/GRTint.c: Rebuilt.
	* mc-boot/GStdIO.c: Rebuilt.
	* mc-boot/GStrCase.c: Rebuilt.
	* mc-boot/GStrIO.c: Rebuilt.
	* mc-boot/GStrLib.c: Rebuilt.
	* mc-boot/GStringConvert.c: Rebuilt.
	* mc-boot/GSysStorage.c: Rebuilt.
	* mc-boot/GTimeString.c: Rebuilt.
	* mc-boot/Galists.c: Rebuilt.
	* mc-boot/Gdecl.c: Rebuilt.
	* mc-boot/Gkeyc.c: Rebuilt.
	* mc-boot/Glists.c: Rebuilt.
	* mc-boot/GmcComment.c: Rebuilt.
	* mc-boot/GmcError.c: Rebuilt.
	* mc-boot/GmcLexBuf.c: Rebuilt.
	* mc-boot/GmcMetaError.c: Rebuilt.
	* mc-boot/GmcOptions.c: Rebuilt.
	* mc-boot/GmcPretty.c: Rebuilt.
	* mc-boot/GmcPrintf.c: Rebuilt.
	* mc-boot/GmcSearch.c: Rebuilt.
	* mc-boot/GmcStack.c: Rebuilt.
	* mc-boot/Gmcp1.c: Rebuilt.
	* mc-boot/Gmcp2.c: Rebuilt.
	* mc-boot/Gmcp3.c: Rebuilt.
	* mc-boot/Gmcp4.c: Rebuilt.
	* mc-boot/Gmcp5.c: Rebuilt.
	* mc-boot/GnameKey.c: Rebuilt.
	* mc-boot/GsymbolKey.c: Rebuilt.
	* mc-boot/Gvarargs.c: Rebuilt.
	* mc-boot/Gwlists.c: Rebuilt.

+++ gcc/m2/mc-boot/GArgs.c
@@ -76,7 +76,7 @@ extern "C" unsigned int Args_GetArg (char *a, unsigned int _a_high, unsigned int
   unsigned int High;
   unsigned int j;
 
-  j = static_cast<unsigned int> (0);
+  j = 0;
   High = _a_high;
   if (i < UnixArgs_ArgC)
     {
