+++ gcc/m2/mc-boot/GStdIO.c
@@ -258,8 +258,8 @@ extern "C" StdIO_ProcRead StdIO_GetCurrentInput (void)
 
 extern "C" void _M2_StdIO_init (__attribute__((unused)) int argc, __attribute__((unused)) char *argv[])
 {
-  StackWPtr = static_cast<unsigned int> (0);
-  StackRPtr = static_cast<unsigned int> (0);
+  StackWPtr = 0;
+  StackRPtr = 0;
   StdIO_PushOutput ((StdIO_ProcWrite) {(StdIO_ProcWrite_t) IO_Write});
   StdIO_PushInput ((StdIO_ProcRead) {(StdIO_ProcRead_t) IO_Read});
 }
